+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/fixtures/TempArticleFixture.java
@@ -1,6 +1,6 @@
 package com.woowacourse.gongseek.acceptance.support.fixtures;
 
-import com.woowacourse.gongseek.article.presentation.dto.TempArticleRequest;
+import com.woowacourse.gongseek.article.presentation.dto.ArticleRequest;
 import com.woowacourse.gongseek.auth.presentation.dto.AccessTokenResponse;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -11,7 +11,7 @@ import org.springframework.http.MediaType;
 public class TempArticleFixture {
 
     public static ExtractableResponse<Response> 임시_게시물을_등록한다(AccessTokenResponse tokenResponse,
-                                                             TempArticleRequest request) {
+                                                             ArticleRequest request) {
         return RestAssured
                 .given().log().all()
                 .header(HttpHeaders.AUTHORIZATION, "Bearer " + tokenResponse.getAccessToken())
